separate c_buffer
